+++ CHANGES.rst
@@ -64,6 +64,8 @@ Usability, bells and whistles
 * The Apply Patches dialog now reports errors when patches fail to apply.
   (`#673 <https://github.com/git-cola/git-cola/issues/673>`_)
 
+* The File menu has a new "Patches" sub-menu with a full set of "git am" Patch actions.
+
 
 Development
 -----------